Add code to ensure no spp are eradicated in APP when setting FNs
I think that results for some of the greedy reserve selectors are being
distorted by eating up the entire landscape searching for species that
have had all of their occupancies removed by adding FNs.
I've now made a change to the way FNs are added so that there is a check
to see if any species has been completely eliminated as a result.
If any have and the new no_empty_spp_allowed flag is set to TRUE, then
at least one of the FNs will be put back so that the spp still have a
minimum number of occurrences.
I have done this in a very quick way just to plug this hole for now and
it can change the total number of FNs generated, so I might want to
change it in the future to behave in a more complex way.  For now though,
I think it will have little bad impact while at the same time, it will
keep some of the greedy algorithms from having artificial blowouts in
their cost errors since they should no longer eat up the entire landscape
looking for non-existant spp occurrences.