Set version number to 3.0
Tzar is flagging the git commit number of btzar but not of bdpg, so it
would take some work to figure out which commit of bdpg is being used in
an experiment, if it's even possible since the nectar image version of
bdpg could be different on every machine if you didn't install/update to
the same version on every machine every time.  I haven't been doing anything
with the version number here in the package DESCRIPTION other than the
initial setup setting the number.  I'm going to start changing the number
at each commit to improve reproducibility in the nectar runs by being able
to find which version/commit generated the bdpg package being used.

I've chosen to set the new version number to 3 instead of 1 or 2 because
the package is a couple of years old now and at one point, I made a release
number of 2 on github when sharing some code with Ascelin and I don't want
to confuse the numbering relative to that.