Add starting_dir arg to all problem creation functions
All of these routines were setting the top directory of the problem to
be the value of parameters$full_output_dir_with_slash.
Need to be able to specify different roots of the problem dir tree in
different situations, particularly when building an ensemble of problems
associated with a particular base problem.