Add starting_dir for repro_...() functions - but possibly not right
These 2 repro functions both call other functions that now need the
starting_dir argument passed to them.  Nearly everything now needs
that argument passed down from the top, but I'm not sure what effect
passing that in would have on these repro functions, so I'm just
setting the start_dir to what they're already using for that.
I don't know if that's the right thing to do and this needs more
attention.  However, these functions are currently not called anywhere,
so it's not a pressing issue to work this out.  I've added issue 59 on
github in the bdpg repository to flag this.