Add first cut at ensemble RS control logic
Added code that recognizes the command to run an ensemble and creates the
RSrun for the ensemble.  It then generates N apparent problems inside the
RSrun's directory to make up the ensemble.  It also runs marxan on each of
those apparent runs.  It also adds logic to turn off any other reserve
selectors when running on these subproblems inside the ensemble.  It still
runs whichever reserve selectors are selected on the problems at the top
level, i.e., problems not generated just to make up the ensemble.
At this point though, it doesn't actually do any aggregation to create
the ensemble solution.
solution vectors