Remove ens_run_dir from numerous ens calls
Many of the ensemble routines have ens_run_dir as an argument at the same
time that they have starting_dir as an argument.  I think they're both
the same thing, but I'm not 100% sure at the moment, so I've just
commented out all of the uses of ens_run_dir with a comment that looks like:
#2018 12 23#.  Once everything is tested, it will be clear whether you
can go ahead and remove all of these lines.  For the moment though, I want
to remember where they are, in case I'm wrong.  It may be that one or both
of starting_dir and ens_run_dir should be set to do_ensemble()'s variable
called base_outdir.