Create empty ens_control_values list as placeholder until values are known
Many calls down the tree from ensemble() require a value for the variable
rs_control_values so it can be echoed to the final csv file of results,
however there are no control values like this defined for the ensemble yet
so it's empty for the moment.  Things like the ensemble size should go in
there and probably something related to whatever contorls the error models
used in the ensemble APP problem building.