Fixed wrap bug: changed num_spp_to_generate to tot_num_spp_after_wrapping
Replacing num_spp_to_generate with tot_num_spp_after_wrapping
because num_spp_to_generate is ambiguous, i.e., is it the number
of new species to add in the wrap or is it the total number of
species in the final wrapped distribution including the original
Xu species?  The correct answer is the total number of species
in the final wrap, but the code that was here doesn't do that.

I supposedly fixed a bug in this back on 2018 11 08 in commit
de06807e "Fix bug in computing num_spp_to_generate for wrap".
Instead of fixing it, I made it calculate the number of new species
to add in the wrap (i.e., excluding the original Xu species),
however, it needs to be the total number of species in the final
wrapped distribution.  This led to fewer species in the wrapped
distribution, which still worked in the sense that it didn't
usuall crash the wrapping code, but it made the fraction of Xu
species in the final distribution slightly off.
However, when the desired_Xu_spp_frac_of_all_spp was 50%, it did
crash the wrapping code because it meant that the lognormal search
was trying to fit a lognormal that had exactly the same number of
total points as points on exactly 2 PUs.

So, I'm changing the name to make it clearer exactly what the
aim of this computation is and changing the computation itself so
that it gets to that aim instead of the incorrect aim that was
introduced in the November "bug fix".  Hopefully, this time it
really IS a fix.  To that end, I'll also add a validation test
after the wrap is built and in that test, make sure that the
desired fraction is met.

  OLD (INCORRECT) CODE that calculated num new spp to add:
      num_spp_to_generate =
          round (Xu_tot_num_spp * (1 - desired_Xu_spp_frac_of_all_spp) /
                                  desired_Xu_spp_frac_of_all_spp)

  CORRECT CODE:
    tot_num_spp_after_wrapping = round (Xu_tot_num_spp / desired_Xu_spp_frac_of_all_spp)