Reinstate passing ensemble dir into save_...() ensemble routines
Had removed passing this dir into various routines because I thought it was
the same as the starting dir, but that was wrong since starting dir was
the top of the whole run.  That meant then ensemble result files (ENS_...)
were getting written out at the top of the run level rather than at the top
of the ensemble level.