Fix bug from no repeatable random seed value for marxan when none is given
The marxan setup code was using the seed of the RSrun as the value to
give to marxan as its random seed, but that value is often NA since no
seed was given to the RSrun.

Here's the warning message from marxan in the output:
WARNING: found bad value for variable RANDSEED. Value ignored

The value for RANDSEED appears in the input.dat file for marxan and
it showed:
RANDSEED NA

So, now I'm just generating a random number every time and using that as
the seed.  That way, it should be repeatable even if the run is rerun
with a specified seed for the bdpg run as a whole.  This is because the
random number generator should still be in the same state at this point.

I've also added a test just before the marxan random seed value is set
and I have it crash if it's NA.  This should never happen, but this makes
sure of that.