Fix missing assignment to dependent_node_IDs (re: github issue #60)
Documented this in bdpg github issue #60.

Added this line to gen_single_bdprob.R:

Xu_bdprob_cor@dependent_node_IDs = PU_spp_pair_info@dependent_node_IDs # Added 2018 12 31 - BTL

The dependent_node_IDs slot already exists in the Xu problem class, but my tests today show that its value has never been getting set. That's because its value IS set in the PU_spp_pair_info object and then every time it's ever needed, it's either passed in to the function of interest or it's derived from the nodes object using get_dependent_node_IDs().

It looks like the slot is completely vestigial and unnecessary and should be removed, however, that would change the Xu problem class and might break reading any old copies of the class from disk.

So, I'm going to just set the value now and assume that it won't hurt anything because it's never referenced in the code. May want to remove the slot in the future.