Fix missing assignment to dependent_node_IDs in wrap (github issue #60 again)
Had same problem in wrap as in commit fbb7a195 for gen_single_bdprob.R,
so I've made a similar change in wrap.  Here though, the PU_spp_pair_info
structure didn't exist like it did for the other case, so I've just
assigned the value here to be the same value as in the base problem that
the wrap problem is wrapping, since that is the correct value by defn.
This is also documented in github issue #60.