Add function to gen 4 simple random variants
This code is different from the original code to generate 4 basic variants
in that its main purpose is to generate a small, simple set of problems
to use in the first paper where it's explaining and demonstrating the
problem generator.  The problems generated here have 1 COR Base, 1 COR Wrap,
1 APP Wrap with random, unmatched FN, FP, and no cost error, and 1 APP Wrap
with random, unmatched FN, FP, and WITH cost error.