Undo previous commit to allow github release at old state
I want to make a github release with the code in rpughly
the state it was in when the bdpg papers' experiments were
finished. However, I recently made a single commit to fix
a bug that I had found while working on paper 1.  I don't
want that fix to be in the release since I want to make
another release with any bug fixes resulting from working

There are ways to roll back a commit using git, but I
have little or no experience with them and everything
I read about them says Danger all over.  So, I'm going
to do the simple, blunt force way of doing it.  I've
just set aside a copy of the gscp_5...R file with the
bug fix in it to recommit later.  With the current
commit, I've gone to github and recovered the gscp_5...R
file from before the bug fix commit and put that back
in here.  I'll commit that now and then go do the
github release.  After that, I'll come back and recommit
the bug fix version.  I'm doing all this because I don't
think that github allows you to do a release by pointing
to a certain commit.  I think that you're only allowed
to point at the head of a specific branch.