Bump version to 3.1.0 to flag end of version used in experiments
I'm about to make a bug fix or two as a result of working
on the bdpg papers, but these fixes won't have been in
the version used to do the experiments.  So, I'm going
to make a github release of 3.1.0 as something close to
what was used in the experiments.  Versions after this
point will include bug fixes.
Even though 3.1.0 is close to what was run during the
experiments, not every experiment was run with the same
code since some bugs and features were handled along the
way.  However, 3.1.0 is as close to the experiment
version as someone could reasonably download right now.