Regenerate documentation & remove unneeded inheritParams uses
Reran the Document command in RStudio's Build/More menu to
regenerate all of the .Rd files that I had deleted to ease
merging of master and ensemble_1 branches.
The command worked but generated some warnings of the form:

Warning messages:
1: Topic 'apply_const_error_to_spp_occupancy_data': no parameters to inherit with @inheritParams
2: Topic 'create_Xu_problem_from_scratch_given_params': no parameters to inherit with @inheritParams
3: Topic 'always_set_new_or_forced_rand_seed': no parameters to inherit with @inheritParams
4: Topic 'get_bdprob_from_rds_file': no parameters to inherit with @inheritParams

These were all functions that included the documentation
line:
    #' @inheritParams std_param_defns
but had all of their arguments were specified locally
so there were no undocumented arguments to inherit.
Deleted the line from each of the functions and that
fixed the problem.
I also had to temporarily comment out some lines in
do_ensemble_analysis_and_output.R that are currently
there to flag what needs to be done next and to purposely
keep the code from running.  They're just text without
comments, so the documentation function broke trying to
parse the file.  After the documentation has run, I have
reverted the file and not included it in the commit.