Repeat rescinded bug fix - adding k__arity exponent to target_num_links_between_2_groups_per_round
This fixes a bug where the arity used was 1 instead of 2.
Found this while working on bdpgtext paper 1.
I originally fixed this back on 2021 01 24, but ended up
backing it out because I needed to have a release of the
bdpg package that most closely approximated the code used
to generate the results used in the bdpg papers.
The version of gscp_5...R that I'm committing now is
the same file that was committed back in January.
I went into the git history in RStudio and found it at
commit 2c395da8.  I had trouble finding it on github
until I did a search there by inserting "hash: 2c395da8"
in the search box.  The full SHA/hash for the commit is
2c395da860751b9f414f121bc02526d76816e602.
The date on it is 2021-01-24-05:12.
The Subject is "Bug fix - adding k__arity exponent to
target_num_links_between_2_groups_per_round".
The Parent is
15a361654cbf19a038f857e70ecc8e9e8fc8147e.
I've put a bunch of notes about this problem of finding
the old commit in the daily log file for May 18, 2021.
There was also issue #61 on the bdpg github repo that
had notes that helped me find the commit.