Temporarily move ensemble code to holding area
I hadn't finished working on this code when I stopped
working on the ensemble branch a while back and I had
left some comments without comment markers to tell me
where to start in again.  That caused the package build
to fail, so I'm just moving the code to a holding area
where it won't bother anything until I can get back to it
and I can leave the naked comments there to tell me where
to start in again.