Fix bug in wrapped landscape size
Wrapping was computing the landscape size using the total
number of nodes in the base problem instead of using the
number of nodes in just the base problem solution when
trying to match the solution_frac_of_landscape parameter.
This made the actual solution fraction of the landscape
slightly smaller than desired.  It was only slightly
smaller because the difference between the size of the
full base problem and the base problem solution is
generally small.