Move ggplot_20_variant_results.R to a new Old_code dir
This file was just part of exploring how to use ggplot
at one time and really isn't part of the package
functionality, but I'm not ready to throw it away yet.
So, I've created an Old_code directory and put it there
for now.  Eventually, it should be removed from this
package.  I'm moving it right now because it causes
many warning messages to come up when I run devtools
Check in RStudio.